+++ src/aiq/agent/tool_calling_agent/agent.py
@@ -25,7 +25,7 @@ from langgraph.prebuilt import ToolNode
 from pydantic import BaseModel
 from pydantic import Field
 
-from aiq.agent.dual_node import AgentDecision
+from aiq.agent.base import AgentDecision
 from aiq.agent.dual_node import DualNodeAgent
 
 logger = logging.getLogger(__name__)
@@ -106,7 +106,7 @@ class ToolCallAgentGraph(DualNodeAgent):
 
     async def build_graph(self):
         try:
-            self.graph = await super()._build_graph(state=ToolCallAgentGraphState)
+            await super()._build_graph(state_schema=ToolCallAgentGraphState)
             logger.info("Tool Calling Agent Graph built and compiled successfully")
             return self.graph
         except Exception as ex:
